requirements.txt
selenium
webdriver-manager
beautifulsoup4
requests
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
python-dateutil